Trying to fix the build check & port cfgs so it's not as messy (#110)
* added cfgs for everything and changed tests

* push some can change

* uart mt fix, also renamed cfg files

* dirty commit 💩

* update yml

* removed cfg files (unnecessary for this set of changes)

* removing unnecessary

* Revert "removed cfg files (unnecessary for this set of changes)"

This reverts commit 8d328bfdff4020acf79ceb6789b8686ab2df362c.

* remove BSP_SOURCES from cfg file bc not necessary anymore

* address review

* address review comments; change relative paths to absolute

* changed some bear stuff; kinda messy but lmk

* had to not use recursion on some variables; changed makefile bear structure again to have move-commands as an explicit target to avoid wonky if statements everywhere

* flash --depends on--> all; supports multiple targets now like make clean all flash without running 3x

* wrote porting guide, was able to not require cfg file and just use linker script to determine valid ports; modified makefile wildcard rule to not try to run default as a target

* updated readme with port guide

* bah

* __weak review fix

* only one bin can exist at a time in the build directory; on build it removes all pre-existing bins, hexs, elfs



Fixed our build checks. Now, we error out if a specific port does not support a feature. The way I implemented this was to put an ifdef in the header and use the error directive there; that way, if the header is every included by anything (meaning we need to use that BSP module) an error will be thrown if the port does not support.

Also, made the makeAllTests parallel, so now it runs two times faster (tested this with two separate commits, took 8m for the non-parallel and 4 minutes for the parallel). Cleaned up the output and made it colorful!

Finally fixed a lot of the build process so that:
make -j doesn't unintentionally build multiple times
bear is not run on targets flash and clean